[develop] Refactoring
